Code design: removed unnecessary base class for Patching and RPU wizards.
Not sure why I had added that since the containing protected method is probably
more relevant as a static method in the Problem class.
