+++ XenAdmin/XenAdmin.csproj
@@ -1096,9 +1096,6 @@
     <Compile Include="Wizards\RollingUpgradeWizard\RollingUpgradeUpgradePage.Designer.cs">
       <DependentUpon>RollingUpgradeUpgradePage.cs</DependentUpon>
     </Compile>
-    <Compile Include="Wizards\UpdateUpgradeWizard.cs">
-      <SubType>Form</SubType>
-    </Compile>
     <Compile Include="Diagnostics\Checks\PoolHasGFS2SR.cs" />
     <Compile Include="Wizards\BootModesControl.cs">
       <SubType>UserControl</SubType>
